Archivist.Backend/pyproject.toml
[project]
name = "archivist-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20250625",
    "pyzmq>=27.1.0",
    "ruff>=0.13.0",
]
